+++ tests/min_tests.py
@@ -101,7 +101,7 @@ def run_testsuit():
         "test_zoom",
         "test_zoom_affine",
         "test_zoomd",
-        "test_compute_occlusion_sensitivity",
+        "test_occlusion_sensitivity",
     ]
     assert sorted(exclude_cases) == sorted(set(exclude_cases)), f"Duplicated items in {exclude_cases}"
