+++ tests/test_img2tensorboard.py
@@ -15,7 +15,7 @@ import numpy as np
 import tensorboard
 import torch
 
-from monai.visualize import make_animated_gif_summary
+from monai.visualise import make_animated_gif_summary
 
 
 class TestImg2Tensorboard(unittest.TestCase):
