+++ tests/test_plot_2d_or_3d_image.py
@@ -17,7 +17,7 @@ import torch
 from parameterized import parameterized
 from torch.utils.tensorboard import SummaryWriter
 
-from monai.visualize import plot_2d_or_3d_image
+from monai.visualise import plot_2d_or_3d_image
 
 TEST_CASE_1 = [(1, 1, 10, 10)]
